Move libMesh-specific code out of m4 files which don't have to be libMesh-specific.
In the long run I'd also like to move the "first check in contrib/"
behavior into a supplied argument, for similar aesthetic reasons and
to make it easier to replace contrib/ software versions with external
software versions at configure time.

In the short run I mostly just want to update petsc.m4 and then swipe
it for another project.