Updated InverseDistanceInterpolation class to use an internal KD-Tree.
This will need to be updated for the parallel case where we want to
allow for source data on remote processors - and I have an idea how to
do that - but I'm leaving that part for future work.
examples/miscellaneous/miscellaneous_ex8 demonstrates usage so far.

not actually interpolating yet.