Disambiguate calculate_norm(vec, H1)
Using system_norm.h instead of a forward declaration in system.h
ensures that the implicit conversion between enum FEMNormType and
class SystemNorm will be considered.

Removing the optional arguments from the single-variable
calculate_norm() ensures that the FEMNormType *won't* be implicitly
converted to an integer and interpreted as a variable index.