Use old file format in reference_elem.data for Pyramid13 and Pyramid14
Either read_ref_elem() and reference_elem.data need to be updated to
read the extra fields in 'contemporary' xda files, or the contents of
reference_elem.data need to conform to the read_ref_elem() implementation.
For now, I've chosen the latter.  This fixes the new subdomains_ex3
example for me.