FEMContext now supports mixed dimension meshes.
This is accomplished by caching FEAbstract* and QBase* for each
of the different dimensions in the mesh, for each FEType. This is
fully backward compatible - all dimension dependent calls have a
default version that defaults preferably to the element dimension,
if guaranteed available, otherwise to the mesh dimension; for single
dimension meshes, there is no difference.